pyproject.toml
[project]
name = "ai_native_textbook"
version = "0.1.0"
description = "AI Native Textbook - Embedding Pipeline for Docusaurus Content"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.9.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "aiohttp>=3.8.0",
    "google-generativeai>=0.8.5",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
